requirements.txt
fastapi
pdfplumber
pytesseract
Pillow
opencv-python-headless
uvicorn